dom/media/test/test_access_control.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=451958
-->
<head>
  <title>Test for Bug 451958</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=451958">Mozilla Bug 451958</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 451958 **/

function run() {
  window.open("http://example.org:80/tests/dom/media/test/file_access_controls.html", "", "width=500,height=500");
}

function done() {
  mediaTestCleanup();
  SimpleTest.finish();
}

addLoadEvent(run);
SimpleTest.waitForExplicitFinish();


window.addEventListener("message", receiveMessage);

function receiveMessage(event)
{
  if (event.origin !== "http://example.org") {
    ok(false, "Received message from wrong domain");
    return;
  }

  if (event.data.done == "true")
    return done();

  ok(event.data.result, event.data.message);
}
</script>
</pre>
</body>
</html>
